- Version bump to 2.1.7. - Fixed: Placing a /* */ comment at the very end of a file without a   terminating newline would not parse reliably. - Added a NULL WallSpriteTile check to R_RenderDecal().
SVN r365 (trunk)

+++ src/c_dispatch.cpp
@@ -927,7 +927,7 @@ FString SubstituteAliasParams (FString &command, FCommandLine &args)
 
 			// Extract the argument number and substitute the corresponding argument.
 			argnum = strtoul (p + 1 + (p[1] == '{'), &start, 10);
-			if ((p[1] != '{' || *start == '}') && argnum < args.argc())
+			if ((p[1] != '{' || *start == '}') && argnum < (unsigned long)args.argc())
 			{
 				buf += args[argnum];
 			}
